+++ examples/lib/stories/input/joystick_advanced.dart
@@ -12,8 +12,7 @@ import 'package:flutter/painting.dart';
 
 import 'joystick_player.dart';
 
-class JoystickAdvancedGame extends FlameGame
-    with HasDraggableComponents, HasTappableComponents {
+class JoystickAdvancedGame extends FlameGame with HasDraggables, HasTappables {
   late final JoystickPlayer player;
   late final JoystickComponent joystick;
   late final TextComponent speedText;